Add MainActivity and HomeFragment
