+++ AdAway/src/main/java/org/adaway/ui/ListsActivity.java
@@ -72,7 +72,7 @@ public class ListsActivity extends AppCompatActivity {
         switch (item.getItemId()) {
             case android.R.id.home:
                 // app icon in Action Bar clicked; go home
-                Intent intent = new Intent(mActivity, BaseActivity.class);
+                Intent intent = new Intent(mActivity, MainActivity.class);
                 intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
                 startActivity(intent);
                 return true;
